chore: move tasks into scripts [swc-854]
