+++ .github/CODEOWNERS
@@ -1,5 +1,1 @@
-cem-react-wrapper.config.js @jianliao
-scripts/cem-plugin-react-wrappers.js @jianliao
-scripts/generate-icon-react-wrapper.js @jianliao
-tasks/build-react.js @jianliao
 *  @adobe/swc-maintainers
