+++ .github/workflows/browser-tests.yml
@@ -93,7 +93,7 @@ jobs:
               uses: actions/github-script@v7
               with:
                   script: |
-                      const { buildTachometerComment } = await import('${{ github.workspace }}/tasks/build-tachometer-comment.js');
+                      const { buildTachometerComment } = await import('${{ github.workspace }}/.github/scripts/build-tachometer-comment.js');
+                      const { commentOrUpdate } = await import('${{ github.workspace }}/.github/scripts/comment-or-update.js');
                       const body = buildTachometerComment();
-                      const { commentOrUpdate } = await import('${{ github.workspace }}/tasks/comment-or-update.js');
                       commentOrUpdate(github, context, '## Tachometer results', body);
