+++ .github/workflows/urls-smoke-test.yml
@@ -39,10 +39,9 @@ jobs:
               uses: actions/github-script@v7
               with:
                   script: |
-                      const { buildPreviewURLComment } = await import('${{ github.workspace }}/tasks/build-preview-urls-comment.js');
+                      const { buildPreviewURLComment, getDocPreviewURL } = await import('${{ github.workspace }}/.github/scripts/build-preview-urls-comment.js');
+                      const { commentOrUpdate } = await import('${{ github.workspace }}/.github/scripts/comment-or-update.js');
                       const body = buildPreviewURLComment(process.env.GITHUB_HEAD_REF);
-                      const { getDocPreviewURL } = await import('${{ github.workspace }}/tasks/build-preview-urls-comment.js');
-                      const { commentOrUpdate } = await import('${{ github.workspace }}/tasks/comment-or-update.js');
                       commentOrUpdate(github, context, '## Branch preview', body);
 
     playwright-smoke-tests:
