+++ packages/icons-ui/package.json
@@ -59,7 +59,7 @@
         "@spectrum-css/ui-icons-s2": "npm:@spectrum-css/ui-icons@2.0.0-s2-foundations.10",
         "case": "^1.6.1",
         "cheerio": "^1.0.0-rc.2",
-        "fast-glob": "^3.2.12",
+        "fast-glob": "^3.3.3",
         "fs": "^0.0.1-security",
         "path": "^0.12.7",
         "prettier": "^3.5.3"
