+++ packages/icons-workflow/package.json
@@ -59,7 +59,7 @@
         "@adobe/spectrum-css-workflow-icons-s2": "npm:@adobe/spectrum-css-workflow-icons@4.1.0",
         "case": "^1.6.1",
         "cheerio": "^1.0.0-rc.2",
-        "fast-glob": "^3.2.12",
+        "fast-glob": "^3.3.3",
         "fs": "^0.0.1-security",
         "path": "^0.12.7",
         "prettier": "^3.5.3"
