+++ scripts/build-css.js
@@ -1,9 +1,10 @@
+#!/usr/bin/env node
+
 /*
-Copyright 2022 Adobe. All rights reserved.
+Copyright 2025 Adobe. All rights reserved.
 This file is licensed to you under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License. You may obtain a copy
 of the License at http://www.apache.org/licenses/LICENSE-2.0
-
 Unless required by applicable law or agreed to in writing, software distributed under
 the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
 OF ANY KIND, either express or implied. See the License for the specific language
@@ -12,9 +13,11 @@ governing permissions and limitations under the License.
 
 import path from 'path';
 import fs from 'fs';
-import { bundleAsync } from 'lightningcss';
 import { fileURLToPath } from 'url';
 import { createRequire } from 'node:module';
+
+import fg from 'fast-glob';
+import { bundleAsync } from 'lightningcss';
 import { stripIndent } from 'common-tags';
 import 'colors';
 
@@ -56,20 +59,13 @@ const wrapCSSResult = (content) => {
     `;
 };
 
-const licensePath = path.resolve(__dirname, '..', 'config', 'license.js');
-let header = '';
-if (fs.existsSync(licensePath)) {
-    header = fs.readFileSync(licensePath, 'utf8');
-    header = header.replace('<%= YEAR %>', new Date().getFullYear());
-}
-
 /**
  * Processes a CSS file using lightningcss, minifies it, and outputs a TypeScript module.
  * The output module includes license headers and wraps the CSS in a template literal.
- * 
+ *
  * @param {string} cssPath - Path to the CSS file to process
  * @returns {Promise<void>} A promise that resolves when processing is complete
- * 
+ *
  */
 export const processCSS = async (cssPath) => {
     return bundleAsync({
@@ -93,14 +89,31 @@ export const processCSS = async (cssPath) => {
         .then(({ code }) => {
             log.success(cssPath.yellow + ' bundled successfully');
 
-            fs.writeFileSync(
-                `${cssPath}.ts`,
-                header + wrapCSSResult(code),
-                'utf-8'
-            );
+            fs.writeFileSync(`${cssPath}.ts`, wrapCSSResult(code), 'utf-8');
         })
         .catch((er) => {
             log.fail(cssPath.yellow + ' failed to bundle');
             console.error(er);
         });
 };
+
+const buildCSS = async () => {
+    const promises = [];
+    for (const cssPath of await fg([
+        './packages/*/src/**/*.css',
+        './tools/*/src/*.css',
+        './tools/*/src/**/*.css',
+    ])) {
+        promises.push(
+            processCSS(cssPath)
+                .then(() => console.log(`Processed ${cssPath}`))
+                .catch((error) =>
+                    console.error(`Error processing ${cssPath}`, error)
+                )
+        );
+    }
+
+    return Promise.all(promises);
+};
+
+await buildCSS();
