+++ scripts/check-cem.js
