+++ scripts/remove-custom-elements-json.js
