+++ scripts/update-version.js
