--- tasks/build-exports-list.js
@@ -1,41 +1,0 @@
-/*
-Copyright 2022 Adobe. All rights reserved.
-This file is licensed to you under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License. You may obtain a copy
-of the License at http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software distributed under
-the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
-OF ANY KIND, either express or implied. See the License for the specific language
-governing permissions and limitations under the License.
-*/
-import path from 'path';
-import fs from 'fs';
-import fg from 'fast-glob';
-
-const build = (packageJSONPath) => {
-    const exportsPath = path.join(packageJSONPath, '..', 'exports.json');
-    if (fs.existsSync(exportsPath)) {
-        return;
-    }
-    const packageJSON = JSON.parse(fs.readFileSync(packageJSONPath, 'utf8'));
-    const exports = packageJSON.exports;
-    if (!exports) {
-        return;
-    }
-    delete exports['.'];
-    delete exports['./package.json'];
-    fs.writeFileSync(exportsPath, JSON.stringify(exports, null, '    '), {
-        encoding: 'utf8',
-    });
-};
-
-const buildExportsList = async () => {
-    for (const packageJSON of await fg(`./**/package.json`, {
-        ignore: ['**/node_modules/**'],
-    })) {
-        build(packageJSON);
-    }
-};
-
-buildExportsList();
